同步SecurityHelper代码
